Rewrite everything to support templates
This is a squash commit of a few months of work, simply because
the code history could never be compiled by anyone, because it
relied heavily on local forks of libraries (like `parse_wiki_text`)
which don’t ever deserve to be published anywhere because they
were never working correctly.

I suppose everything up to now could be squashed but let’s just
say that humility requires allowing at least some of your nasty
mistakes to be present in public.