Fix infinite backtracking due to byte/character index confusion
In the case where `ustring.gmatch` matches exactly at the end of
a string with some Unicode characters in it, using the character
index instead of the string byte length will cause backtracking.
Apparently there is no test case for this. 🥴

Really these functions should all just be using character indexes
and there should be a newtype to prevent confusion, but, lol.