Use table layout in source viewer
Turns out grids are limited to 10,000 rows in Gecko, lol. Well,
that’s fine, there is actually no reason to not use the table
layout model when all the text line heights are the same, so the
baselines all line up just fine, even with `vertical-align: top`.
(Yes, that is a pretty bad reason to use grid layout for a table.
Don’t @-me.)