Add some missing i18n separators
It is totally unclear at the moment what the message source is
for all of these keys. At least, searching the wikimedia org for
these keys does not show any single file containing all of them.
So until such a master list can be found, it will continue to be
(like everything) a game of trying to add things that seem
relevant as pages are subtly broken.