Emit parameter fallbacks more correctly, and cache missing articles
I love spending hours chasing what looks like an awful edge case
defect only to find out that someone wrote garbage which is
deliberately designed around the failure condition. Success is
failure, the truth is a lie, freedom is slavery. How many gaslights
are there, Picard?

Anyway, this improves performance on pages that use
'Module:CountryData'. Whilst it has a cache of its own, it also
just goes ahead and ignores that whenever `getcachedtable` is
called with parameters, which for at least 'Module:Flagg' is
*every call*.