Embrace the reality that template output must be treated as stringy
In the beginning, I falsely believed that Wikitext was a format
whereby it would be possible to tokenise templates as they were
expanded and then pass the new tokens up to the root document for
processing, and the architecture of the renderer was set up to
support that very obvious and efficient design. Oh, what a fool
I was, a pitiable and naïve fool! Imagining that any software
programmer would spend a hot second thinking about the formal
grammar of a new structured text format they’re inventing when
regular expressions are *just sitting there*.

“Pick me!” whispers the Perl-compatible regular expression engine
to every programmer who’s ever glimpsed its man page. “I’m so easy
to use! You can solve all your text parsing needs with me! Ignore
Chomsky, that guy sucks!” Does the regular expression engine do
this because it is evil? Because it is lonely? Perhaps. But perhaps
such outcomes are expected when a good tool has been abused. And
regular expressions certainly have been abused. One cannot simply
drop a type-3 automaton into a type-1 world without warping its
sense of what it can, and cannot, expect from itself.

And what makes otherwise intelligent and very handsome engineers
grow so hot and weak at the knees for regular expressions every
time they see what appears to be some structured text, no matter
how irregular or context-sensitive? Why do they insist that all
strings are just begging to be plucked apart by some finite
automaton? Is it that they feel pity for the regular expression
engine because they can see how incapable it is and want to make
it feel more powerful? Or is it the opposite, a blind faith that
any regex engine can be made to cleave text into pieces, no
matter its grammatical complexity? “Maybe this time the regex
*will* be able to parse HTML, if I only use it a *little* bit,”
the developer thinks to themselves. “Maybe this time there won’t
be any consequences at all”, they rationalise. A memory of the
gospel of bobince 1:17 from the good book (Stack Overflow)
flickers through their mind, but only for a moment, as their
stupid lazy fingers start to type the most dangerous letters:

p.
r.
e.
g.

The incantation is complete. They are locked in. We are all locked
in. You used a regular expression. Now there are two problems.
Damn you.

So anyway, the only way to get a fully correct Wikitext parse is by
wasting energy tokenising every template, then throwing away most
of that work and starting over, because the output of a template is
a string, not a serialised Wikitext token tree, and what tokens
those strings represent depend on the entire sequence of template
expansions being concatenated together as strings first. There’s
nothing anyone can do about it, except for that one lady with the
time machine, and she’s too drunk to go back in time and fix it
(and, let’s be honest, there are clearly more important things that
time machine lady needs to focus on going back and fixing). So
we’re stuck, and might as well not write a uselessly complex token
emitter when it’s all getting glued together as text later.