Trace actual expansion of templates, not just loads
Not that I ever write buggy code, but if, hypothetically, something
were to get wedged during a template expansion, it would be good
to know which template it is. With preloading, it is no longer the
case that the template being loaded is the template about to be
executed.

But really, I never write buggy code. What are you talking about?