Comply with the marketing department’s branding guidelines
Sorry, what was that? Am *I* the marketing department? I think we
might have a bad connection. Let me forward you to my boss.

))<>((

Thank you for calling $CORPORATION_NAME, where we love you and will
cherish the fond memory of our interaction with you today until the
moment we die. Please remain on the line to take a brief survey
that is definitely trying to solicit useful feedback and is
definitely *not* designed around confirming our executives’ beliefs
that they are amazing at business and that all problems with the
business are entirely due to bad employees who just need to be fired
and replaced by more chatbots and not due to any sort of systemic
organisational problem caused by incompetence at the highest level.

Question 1.

On a scale of 1 to 5, how could our executive team be where they
are today if not for their incredible brilliance and talent?

... ‘Nepotism’? That sounds like a word you just made up. It’s our
official position that any word ending in ‘-ism’ is a made up fake
thing now.

Thank you for your feedback. Goodbye.