Strip category whitespace without breaking other things
Turns out that empty `<nowiki>` strip markers are not just there
for fun, they are the blessed way to prevent parser functions and
such from trimming whitespace. So it isn’t possible to just not
emit them. But then they get in the way of trimming the category
whitespace where Wikitext really, really, really wants that to be
gone, even if a strip marker is in the way. So now it is, and the
world is spinning out of control, and I accidentally just cleared
my terminal so I can’t see where I am in this commit message any
more. This is fine.

Also, this finally puts the mechanism in place that parser
functions actually need to work correctly in the presence of
strip markers, so that’s good I guess.