Use a whitelist for additional styling hooks
Since CSS has to be written by someone anyway, there is no reason
to add extra style hooks *everywhere*, given how problematic this
has turned out to be due to various hypersensitivities among the
modules and their unwillingness to go to therapy to address it.

This does make it a bit harder at a glance to identify where
potential styling hooks may come from, but without the ability to
put markers on everything due, that is already somewhat of a
problem anyway.